Start breaking up the monolithic `linter.rb`; fix a bug
